hw#11 refactoring v1.9.0

+++ hw11_refactoring/src/domain/comment-service.ts
@@ -1,5 +1,4 @@
 import {CommentsRepository} from "../repositories/comments-repository";
-import {CommentsDbType} from "../types/comment-types";
 
 class CommentService {
     commentRepo: CommentsRepository
@@ -27,14 +26,21 @@ class CommentService {
         return await this.commentRepo.updateComment(commentId, content)
     }
 
-    async deleteComment(commentId: string): Promise<boolean> {
+    async deleteComment(commentId: string, userId:string): Promise<boolean | string> {
+        const resault = await this.commentRepo.findCommentsById(commentId)
+        if (!resault) {
+            return false
+        }
+        if(resault.userId !== userId){
+            return '403'
+        }
         return await this.commentRepo.deleteComment(commentId)
     }
 
     //check below
-    async getDbCommentById(commentId: string): Promise<CommentsDbType | null> {
+    /*async getDbCommentById(commentId: string): Promise<CommentsDbType | null> {
         return await this.commentRepo.getCommentById(commentId)
-    }
+    }*/
 }
 
 export const commentService = new CommentService()
